requirements.txt
Flask==3.0.0
mysql-connector-python==9.0.0
Werkzeug==3.0.1
pillow==10.2.0
google-generativeai==0.5.2
pandas==2.2.0
requests==2.31.0
fpdf==1.7.2
pyotp==2.8.0
qrcode==7.4.2
gunicorn==21.2.0
python-dotenv==1.0.1
